trackwise-backend/ec2_embed/requirements.txt
boto3
faiss-cpu
sentence-transformers
